Use macro to generalize behaviour.
